+++ apps/api/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @charlybot/api
 
+## 2.4.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @charlybot/shared@2.5.0
+
 ## 2.4.1
 
 ### Patch Changes